requirements.txt
Django==1.10.7
djangorestframework==3.6.2
django-filter==1.0.2